pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-recognizer"
version = "0.1.0"
description = "OCR Recognizer MVP with FastAPI, primary OCR, fallback, and a simple web UI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "pypdf>=4.3.1,<5.0.0",
  "reportlab>=4.2.2,<5.0.0",
]

[project.optional-dependencies]
ocr = [
  "ocrmypdf>=16.4.0,<17.0.0",
]
ocr_fallback = [
  "easyocr>=1.7.1,<2.0.0",
  "pdf2image>=1.17.0,<2.0.0",
  "opencv-python-headless>=4.10.0.84,<5.0.0",
  "pillow>=10.4.0,<13.0.0",
]
dev = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
